requirements.txt
python-dotenv
pandas 
numpy
pymongo
pymongo[srv]==4.7
certifi
#-e .